Glenan dataset

Maerl bed data set to illustrate Ecological Quality Assessment (EQA)

Format

Glenan is an object of class data.frame composed of 32 observations and 252 variables.

Abundance.x

Abundance (number of individuals) of each taxon x

Surveys

Indicates different Maerl bed surveys.

Treatment

Combinations of fishing dredges and pressure levels. 'CTRL' stands for control. Fishing dredges are:

(1) a clam dredge (CD), 70 to 90 kg, 1.5 m wide, 40 teeth of 11 cm each;

(2) a queen scallop dredge (QSD), 120 kg,1.8 m wide, with a blade;

(3) a king scallop dredge (KSD), 190 kg, 1.8 m wide, 18 teeth of 10 cm each every 9 cm.

Details

Experimental data set built by Tauran et al. (2020) to study the impact of fishing dredges and varying fishing pressures on maerl beds, in the bay of Brest (Brittany, France).

References

Tauran, A., Dubreuil, J., Guyonnet, B., Grall, J., 2020. Impact of fishing gears and fishing intensities on maerl beds: An experimental approach. Journal of Experimental Marine Biology and Ecology 533, 151472. https://doi.org/10.1016/j.jembe.2020.151472

See also