heatmapdata dataset

Espinasse et al. (2020) tested the application of isoscapes modelled from satellite data to the description of secondary production in the Northeast pacific. The output model fits in a 0.25° x 0.25° spatial grid covering the region spanning from 46 to 62°N and from 195 to 235°E and supporting δ 13C and δ 15N isoscapes from 1998 to 2017.

Format

heatmapdata is an object of class dataframe composed of 9206 observations of 9 variables.

Latitude

Latitude coordinate of the station, in degrees

Longitude

Longitude coordinate of the station, in degrees

d13C

δ13C modelled value

d15N

δ15N modelled value

station

Station ID

Years

Period corresponding to the calculation of trajectory metrics

Angles

Angle α (i.e direction) in the stable isotope space

Lengths

Net change values (i.e direction) in the stable isotope space

Angles2

Angle alpha values (i.e direction) in the stable isotope space transformed for a potential use with function geom_spoke

Details

This data sets is composed of trajectory metrics calculated by Sturbois et al. (2021) for all stations within all inter-annual consecutive periods between 1998 and 2017 calculated from the whole data set of Espinasse et al. (2020) for a 1° x 1° spatial grid.

References

Espinasse, B., Hunt, B.P.V., Batten, S.D., Pakhomov, E.A., 2020. Defining isoscapes in the Northeast Pacific as an index of ocean productivity. Global Ecol Biogeogr 29, 246–261.

See also

Author

Espinasse, B., Hunt, B.P.V., Batten, S.D., Pakhomov, E.A.